requirements.txt
argon2-cffi==18.1.0
bcrypt==3.1.4
cffi==1.11.5
Django==2.0.6
django-bootstrap4==0.0.6
django-filter==1.1.0
Markdown==2.6.11
Pillow==5.0.0
pycparser==2.18
Pygments==2.2.0
pytz==2018.4
six==1.11.0

